+++ app/Http/Controllers/RequisitionController.php
@@ -37,7 +37,8 @@ class RequisitionController extends Controller
             'quotation',
             'company:id,name,logo',
             'machines:id,machine_model_id',
-            'machines.model:id,name'
+            'machines.model:id,name',
+            'user'
         )->whereType('purchase_request')->latest();
 
         $requisitions = $requisitions->has('partItems');
@@ -60,7 +61,7 @@ class RequisitionController extends Controller
                 $requisitions = $requisitions->where('status', $request->status);
             });
 
-            if ($request->comPartReq)
+        if ($request->comPartReq)
             $requisitions = $requisitions->where('is_company', $request->comPartReq);
 
         //Check if request wants all data of the requisitions
@@ -262,7 +263,8 @@ class RequisitionController extends Controller
             'partItems.part.aliases',
             'partItems.part.stocks' => function ($q) {
                 $q->where('unit_value', '>', 0);
-            }
+            },
+            'user'
         ]);
 
         return RequisitionResource::make($requisition);
@@ -301,9 +303,14 @@ class RequisitionController extends Controller
      */
     public function destroy($id)
     {
-        $requisition = Requisition::find($id)->delete();
-        PartItem::where('model_id',$id)->delete();
-            return message('Requisition deleted successfully');
+        $requisition = Requisition::find($id);
+        if ($requisition) {
+            $requisition->delete();
+            PartItem::where('model_id', $id)->delete();
+            return message('Requisition deleted successfully', 201);
+        } else {
+            return message('Requisition Not Found', 422);
+        }
     }
 
     public function storeClientReqisition(Request $request)
